// app/web/src/lib/theme/useTheme.d.ts
export declare const useTheme: () => {
    colors: {
        primary: string;
        secondary: string;
        background: string;
        text: string;
        error: string;
        warning: string;
        success: string;
    };
    spacing: {
        xs: string;
        sm: string;
        md: string;
        lg: string;
        xl: string;
    };
    borderRadius: string;
    shadows: string[];
};